BREAKING: Rename vault to synapse; Extract modules into separate files
