+++ example/index.js
@@ -1,18 +1,18 @@
 import React, { Fragment } from 'react';
 import { render } from 'react-dom';
-import { Provider as VaultProvider } from '../src/synaptik';
+import { Provider} from '../src';
 import * as stores from './stores';
 import TodoList from './components/todo-list';
 import WeatherForecast from './components/weather-forecast';
 import logger from '../logger';
 
 const App = () => (
-  <VaultProvider stores={stores} logger={logger}>
+  <Provider stores={stores} logger={logger}>
     <Fragment>
       <TodoList />
       <WeatherForecast zipCode={94133} />
     </Fragment>
-  </VaultProvider>
+  </Provider>
 );
 
 render(<App />, window.root);
