+++ src/index.js
@@ -1,0 +1,14 @@
+import Connect, { connect } from './Connect';
+import Provider from './Provider';
+import Store from './Store';
+import useSynapse from './useSynapse';
+import Synapse from './Synapse';
+
+export {
+  Synapse,
+  Provider,
+  Connect,
+  connect,
+  Store,
+  useSynapse,
+}
